feat(phi-plugin): 移植 `user` 模块 & 优化命令传参 & 优化性能
- 新增 `user` 模块的功能，包括查询用户数据、统计信息和成绩列表
- 实现了 `data`、`info`、`lvscore` 和 `list` 四个子命令
- 优化了数据处理和展示逻辑，提高了模块的可用性和准确性
- 优化列表删除性能
- 允许命令传参忽略空格
- 同步上游更改

+++ phi-plugin/model/send.py
@@ -30,7 +30,7 @@ class send:
         return recepit
 
     @classmethod
-    async def getsaveResult(cls, session: Uninfo, ver: int | None = None, send=True):
+    async def getsaveResult(cls, session: Uninfo, ver: float | None = None, send=True):
         """
         检查存档部分
